+++ packages/web/src/client.ts
@@ -6,7 +6,7 @@ import { ConcordiumGRPCClient } from '@concordium/common-sdk';
 
 /**
  * Initialize a gRPC client for a specific concordium node.
- * @param address the ip address of the node, e.g. 127.0.0.1
+ * @param address the ip address of the node, e.g. http://127.0.0.1
  * @param port the port to use when econnecting to the node
  * @param options optional options for the grpc-web transport
  */